optimize import and remove dead code
